Make Host aware of its working dir and droid cmd
Add methods to AbleInterface which allow setting and getting of the
directory from where remote commands are executed and the prefix of the
droid command (e.g. "php droid.phar"). Implement these methods in
AbleTrait.

SynchroniserInterface implementations are to make hosts aware of the
droid command prefix after successful sync.