Host attr in Task args; Add Task.host_filter
Task arguments may now refer to attributes of the current remote host,
e.g. `{{{ host.private_ip }}}`, `{{{ host.variables.role }}}`, etc.

The set of hosts on which a Task is run can be manipulated setting the
Task.host_filter property to an expression which evaluates to a boolean.
The expression is evaluated for each of Task.hosts or Target.hosts and
the Task is executed on a host if the expression evaluates to true. The
expression syntax is that of the Symfony Expression component. For
example: `host.variables['role'] == 'slave'`. This feature depends on
droid/droid-model >= 1.0.3.

Additionally, TaskRunner has been refactored to reduce complexity and
ease testing; test coverage has been improved.